Modularize test utilities
